Add juju client interface to help improve testability
